refactor(Pagination): split into 2 classes
Separate the pagination class in 2 specific classes to clarify
the purpose.
